requirements.txt
django
south
djangorestframework
